// 00. Mission 6/nModule_F14/ph-um-client/src/schemas/createAcademicSemesterSchema.ts
import { z } from "zod";

export const createAcademicSemesterSchema = z.object({
  name: z.string({ required_error: "Name is required" }),
  year: z.string({ required_error: "Year is required" }),
  startMonth: z.string({ required_error: "Start month is required" }),
  endMonth: z.string({ required_error: "End month is required" }),
});